refreshAccessToken

+++ src/utils/cloudinary.js
@@ -1,4 +1,5 @@
 import {v2 as cloudinary} from "cloudinary"
+
 import fs from "fs"
 
 
@@ -7,25 +8,33 @@ cloudinary.config({
   api_key: process.env.CLOUDINARY_API_KEY, 
   api_secret: process.env.CLOUDINARY_API_SECRET 
 });
+// console.log("Cloudinary Config:", cloudinary.config());
+
 
 const uploadOnCloudinary = async (localFilePath) => {
     try {
         if (!localFilePath) return null
         //upload the file on cloudinary
+        // console.log("Uploading file to Cloudinary:", localFilePath);
+
         const response = await cloudinary.uploader.upload(localFilePath, {
-            resource_type: "auto"
-        })
+            resource_type: "auto",
+          })
         // file has been uploaded successfull
-        //console.log("file is uploaded on cloudinary ", response.url);
-        fs.unlinkSync(localFilePath)
+        //pela niche ni line comment mari ti atle public/temp ma show thatu tu avatar file and coverImage file hve ola ye on kri atle me pn kri nakhi
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath); // Remove the locally saved temporary file
+        }
         return response;
 
     } catch (error) {
-        fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
+
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }        
         return null;
     }
 }
-
 
 
 export {uploadOnCloudinary}
